refactor: ArgumentResolver 내 MemberService 사용 로직 삭제
- AuthService만 사용하도록 수정
- DB 접근 로직 포함하지 않도록 수정
