+++ src/main/java/nextstep/common/LoginMember.java
@@ -1,0 +1,13 @@
+package nextstep.common;
+
+public class LoginMember {
+    private final Long id;
+
+    public LoginMember(Long id) {
+        this.id = id;
+    }
+
+    public Long getId() {
+        return id;
+    }
+}
